+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ## v4.1.14 ##
   * Allow specifying FetchLen and LOBFetchLen.
+  * Add Rset.ColumnIndex helper function.
 
 ## v4.1.13 ##
   * Fix bindStirngSlice error of nullInds reuse without reset.